+++ Changelog.md
@@ -1,4 +1,9 @@
-## 4.3.1 - Unreleased
+## 4.4.0 Unreleased
+
+### Added
+
+-   J-Link devices ONB FW can be automatically update when programming
+-   Button to Manually update OB Firmware of a device
 
 ### Changed